refactor file structure into game and renderer
The game folder will contain all functionality, and will not be bound to pyglet at all. This will allow me to reuse it for the gym.
The renderer doesn't have this constraint, but mustn't have any game function code in it, it should just represent the game's state to the screen.