Fix not training agent on temporal datasets
We weren't giving it multiple frames to train over, so it was acting like a regular feedforward network and the recursiveness was just making it more difficult to train.
I've removed the handrolled memory and am instead training every frame, based on what the agent has perceived before (stateful=True). I reset the state of the recursive layers every reset, so each life it forgets the short term memory and starts anew (but keeps its network weights)